Moving KeyValueHandler API into the Metadata class in preparation for moving MetadataKey into the extracted "public api" library for Flogger.
Note: The also removes the nullability for the value, since MetadataKey doesn't require that (this came from the fact that the interface was originally used for both Metadata values and Tags, which can have null values).

RELNOTES=Moved internal KeyValueHandler class into MetadataKey class.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=328165390
